requirements.txt
aiogram>=3.2.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.28.0